+++ flowapi/app/run_query.py
@@ -29,26 +29,29 @@ def check_claims(claim_type):
         @wraps(func)
         @jwt_required
         async def wrapper(*args, **kwargs):
-            json = await request.json
-            query_kind = "NA" if json is None else json.get("query_kind", "NA")
-            log_elements = [
-                "",
-                f"{func.__name__.upper()}",
-                query_kind.upper(),
-                get_jwt_identity(),
-                request.headers.get("Remote-Addr"),
-                kwargs.get("query_id", "NA"),
-            ]
-
-            current_app.query_run_logger.info(":".join(log_elements))
+            json_payload = await request.json
+            current_app.access_logger.info(
+                "AUTHENTICATED",
+                request_id=request.request_id,
+                route=request.path,
+                user=get_jwt_identity(),
+                src_ip=request.headers.get("Remote-Addr"),
+                json_payload=json_payload,
+            )
+            query_kind = (
+                "NA" if json_payload is None else json_payload.get("query_kind", "NA")
+            )
             try:  # Cross-check the query kind with the backend
                 request.socket.send_json(
-                    {"action": "get_query_kind", "query_id": kwargs["query_id"]}
+                    {
+                        "request_id": request.request_id,
+                        "action": "get_query_kind",
+                        "query_id": kwargs["query_id"],
+                    }
                 )
                 message = await request.socket.recv_json()
                 if "query_kind" in message:
                     query_kind = message["query_kind"]
-                    log_elements[2] = query_kind.upper()
                 else:
                     return jsonify({}), 404
             except KeyError:
@@ -64,13 +67,26 @@ def check_claims(claim_type):
                     )
 
             claims = get_jwt_claims().get(query_kind, {})
+            log_dict = dict(
+                request_id=request.request_id,
+                query_kind=query_kind.upper(),
+                route=request.path,
+                user=get_jwt_identity(),
+                src_ip=request.headers.get("Remote-Addr"),
+                json_payload=json_payload,
+                query_id=kwargs.get("query_id", "NA"),
+                claims=claims,
+            )
+
+            current_app.query_run_logger.info("Received", **log_dict)
             endpoint_claims = claims.get("permissions", {})
             spatial_claims = claims.get("spatial_aggregation", {})
             if (claim_type not in endpoint_claims) or (
                 endpoint_claims[claim_type] == False
             ):  # Check access claims
-                log_elements[0] = "UNAUTHORIZED"
-                current_app.query_run_logger.error(":".join(log_elements))
+                current_app.query_run_logger.error(
+                    "CLAIM_TYPE_NOT_ALLOWED_BY_TOKEN", **log_dict
+                )
                 return (
                     jsonify(
                         {
@@ -82,7 +98,11 @@ def check_claims(claim_type):
                 )
             elif claim_type == "get_result":  # Check spatial aggregation claims
                 request.socket.send_json(
-                    {"action": "get_params", "query_id": kwargs["query_id"]}
+                    {
+                        "request_id": request.request_id,
+                        "action": "get_params",
+                        "query_id": kwargs["query_id"],
+                    }
                 )
                 message = await request.socket.recv_json()
                 if "params" not in message:
@@ -100,8 +120,9 @@ def check_claims(claim_type):
                         500,
                     )
                 if aggregation_unit not in spatial_claims:
-                    log_elements[0] = "UNAUTHORIZED"
-                    current_app.query_run_logger.error(":".join(log_elements))
+                    current_app.query_run_logger.error(
+                        "SPATIAL_AGGREGATION_LEVEL_NOT_ALLOWED_BY_TOKEN", **log_dict
+                    )
                     return (
                         jsonify(
                             {
@@ -116,7 +137,7 @@ def check_claims(claim_type):
                     pass
             else:
                 pass
-
+            current_app.query_run_logger.info("Authorised", **log_dict)
             return await func(*args, **kwargs)
 
         return wrapper
@@ -130,6 +151,7 @@ async def run_query():
     json_data = await request.json
     request.socket.send_json(
         {
+            "request_id": request.request_id,
             "action": "run_query",
             "query_kind": json_data["query_kind"],
             "params": json_data["params"],
@@ -152,7 +174,9 @@ async def run_query():
 @blueprint.route("/poll/<query_id>")
 @check_claims("poll")
 async def poll_query(query_id):
-    request.socket.send_json({"action": "poll", "query_id": query_id})
+    request.socket.send_json(
+        {"request_id": request.request_id, "action": "poll", "query_id": query_id}
+    )
     message = await request.socket.recv_json()
 
     if message["status"] == "done":
@@ -170,7 +194,9 @@ async def poll_query(query_id):
 @blueprint.route("/get/<query_id>")
 @check_claims("get_result")
 async def get_query(query_id):
-    request.socket.send_json({"action": "get_sql", "query_id": query_id})
+    request.socket.send_json(
+        {"request_id": request.request_id, "action": "get_sql", "query_id": query_id}
+    )
     message = await request.socket.recv_json()
     current_app.logger.debug(f"Got message: {message}")
     if message["status"] == "done":