+++ flowapi/tests/unit/test_access_control.py
@@ -1,11 +1,12 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
+import json
 import os
+
 import pytest
 from asynctest import return_once
-from .utils import query_kinds, aggregation_types
+from .utils import query_kinds
 
 
 @pytest.mark.asyncio
@@ -21,10 +22,14 @@ async def test_protected_get_routes(route, app):
     route: str
         Route to test
     """
-    client, db, log_dir = app
+    client, db, log_dir, app = app
 
     response = await client.get(route)
     assert 401 == response.status_code
+    with open(os.path.join(log_dir, "flowkit-access.log")) as log_file:
+        log_lines = log_file.readlines()
+    assert 1 == len(log_lines)
+    assert "UNAUTHORISED" in log_lines[0]
 
 
 @pytest.mark.asyncio
@@ -36,7 +41,7 @@ async def test_granular_run_access(
     Test that tokens grant granular access to running queries.
 
     """
-    client, db, log_dir = app
+    client, db, log_dir, app = app
     token = access_token_builder({query_kind: {"permissions": {"run": True}}})
     expected_responses = dict.fromkeys(query_kinds, 401)
     expected_responses[query_kind] = 202
@@ -61,7 +66,7 @@ async def test_granular_poll_access(
     Test that tokens grant granular access to checking query status.
 
     """
-    client, db, log_dir = app
+    client, db, log_dir, app = app
     token = access_token_builder({query_kind: {"permissions": {"poll": True}}})
     expected_responses = dict.fromkeys(query_kinds, 401)
     expected_responses[query_kind] = 303
@@ -89,7 +94,7 @@ async def test_granular_json_access(
     Test that tokens grant granular access to query output.
 
     """
-    client, db, log_dir = app
+    client, db, log_dir, app = app
     token = access_token_builder(
         {
             query_kind: {
@@ -131,7 +136,7 @@ async def test_no_result_access_without_both_claims(
     Test that tokens grant granular access to query output.
 
     """
-    client, db, log_dir = app
+    client, db, log_dir, app = app
     token = access_token_builder({"DUMMY_QUERY_KIND": claims})
     dummy_zmq_server.side_effect = (
         {"id": 10, "query_kind": "DUMMY_QUERY_KIND"},
@@ -156,7 +161,7 @@ async def test_access_logs_gets(
     Test that access logs are written for attempted unauthorized access to 'poll' and get' routes.
 
     """
-    client, db, log_dir = app
+    client, db, log_dir, app = app
     token = access_token_builder({query_kind: {"permissions": {}}})
     dummy_zmq_server.return_value = {"id": 0, "query_kind": "modal_location"}
     response = await client.get(
@@ -168,10 +173,13 @@ async def test_access_logs_gets(
     with open(os.path.join(log_dir, "query-runs.log")) as log_file:
         log_lines = log_file.readlines()
     assert 2 == len(log_lines)
-    assert query_kind.upper() in log_lines[0]
-    assert "UNAUTHORIZED" in log_lines[1]
+    assert "MODAL_LOCATION" == json.loads(log_lines[0])["query_kind"]
+    assert "CLAIM_TYPE_NOT_ALLOWED_BY_TOKEN" in log_lines[1]
     assert "test" in log_lines[0]
     assert "test" in log_lines[1]
+    assert (
+        json.loads(log_lines[0])["request_id"] == json.loads(log_lines[1])["request_id"]
+    )
 
 
 @pytest.mark.asyncio
@@ -183,7 +191,7 @@ async def test_access_logs_post(
     Test that access logs are written for attempted unauthorized access to 'run' route.
 
     """
-    client, db, log_dir = app
+    client, db, log_dir, app = app
     token = access_token_builder({query_kind: {"permissions": {}}})
     response = await client.post(
         f"/api/0/run",
@@ -195,6 +203,6 @@ async def test_access_logs_post(
         log_lines = log_file.readlines()
     assert 2 == len(log_lines)
     assert query_kind.upper() in log_lines[0]
-    assert "UNAUTHORIZED" in log_lines[1]
+    assert "CLAIM_TYPE_NOT_ALLOWED_BY_TOKEN" in log_lines[1]
     assert "test" in log_lines[0]
     assert "test" in log_lines[1]